Separate client lib into smaller modules
