+++ lib/client/isObject.js
@@ -1,0 +1,13 @@
+/**
+ * Check if `obj` is an object.
+ *
+ * @param {Object} obj
+ * @return {Boolean}
+ * @api private
+ */
+
+function isObject(obj) {
+  return obj === Object(obj);
+}
+
+module.exports = isObject;
